requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.1.0
PyJWT==2.10.1
python-socketio==5.10.0
python-engineio==4.8.0
aiofiles==23.2.1
